[enhancement](be_metrics) update scan bytes metric in file_scanner. (#53729)
Problem Summary:

The external part is implemented according to the framework defined by
the unified audit log, Doris metrics, and scanbytes and scan rows in
Profile in #52232.

However,
- Scan bytes in the external table currently represents the bytes
counted by **the top-level File Reader** called by the scan reader
layer.
- Scan rows represents the number of scanned rows of the underlying
storage. The number of scanned rows of parquet/orc does not include the
number of rows of skipped page/rowgroup. **Note: However, there is still
a problem that `jni_reader` has not yet implemented the number of rows
that only contain the storage to be read.**